backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0
alembic==1.14.0
redis==5.2.0
structlog==24.4.0
python-multipart==0.0.18
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
minio==7.2.10
anthropic==0.40.0
gitpython==3.1.43
pymysql==1.1.1
pyodbc==5.2.0
cx-oracle==8.3.0
pytest==8.3.4
pytest-asyncio==0.24.0
ruff==0.8.4
tree-sitter==0.23.2
tree-sitter-languages==1.10.2
